+++ src-tauri/src/websocket/client.rs
@@ -1,7 +1,5 @@
-use crate::{
-    settings::{get_settings, Settings},
-    websocket::server::WebsocketRequest,
-};
+use super::WebsocketRequest;
+use crate::settings::{get_settings, Settings};
 use futures_util::SinkExt;
 use log::{debug, info};
 use serde_json::json;
